# spring/docs/openapi.yaml
openapi: "3.0.1"
info:
  title: "Spring部分的Swagger文档"
  description: "这是项目的Spring部分的Swagger文档"
  version: "1.0.0"
...
paths:
...
  /comments:
    get:
      tags:
      - "评论模块"
      summary: "获取普通评论信息"
      description: "分页获取普通评论信息列表，并支持用户名和文章标题模糊查询"
      operationId: "listComments"
      parameters:
      - name: "page"
        in: "query"
        description: "页码"
        required: false
        schema:
          type: "string"
          description: "页码"
          example: 1
        example: 1
      - name: "size"
        in: "query"
        description: "每页条数"
        required: false
        schema:
          type: "string"
          description: "每页条数"
          example: 10
        example: 10
      - name: "username"
        in: "query"
        description: "用户名（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "用户名（模糊匹配）"
      - name: "article_title"
        in: "query"
        description: "文章标题（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "文章标题（模糊匹配）"
      - name: "content"
        in: "query"
        description: "评论内容（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "评论内容（模糊匹配）"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOCommentsVO"
    put:
      tags:
      - "评论模块"
      summary: "修改评论"
      description: "通过请求体修改评论信息"
      operationId: "updateComment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentUpdateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
    post:
      tags:
      - "评论模块"
      summary: "新增评论"
      description: "通过请求体创建评论信息"
      operationId: "createComment"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentCreateDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultLong"
...
  /comments/statistics/delete-ai/{articleId}:
    post:
      tags:
      - "评论模块"
      summary: "删除文章AI评论（内部）"
      description: "删除文章的AI评论，供内部服务远程调用"
      operationId: "deleteAiCommentsByArticleId"
      parameters:
      - name: "article_id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /comments/scores/batch:
    post:
      tags:
      - "评论模块"
      summary: "批量查询评论评分（内部）"
      description: "根据文章ID列表批量查询评论评分，按角色分组，供内部服务远程调用"
      operationId: "getCommentScoresByArticleIds"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIdsDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListArticleCommentScoresVO"
  /comments/internal/create:
    post:
      tags:
      - "评论模块"
      summary: "创建评论（内部）"
      description: "创建评论，供内部服务远程调用"
      operationId: "createCommentInternal"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Comments"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /comments/batch:
    post:
      tags:
      - "评论模块"
      summary: "批量查询评论（内部）"
      description: "根据ID列表批量查询评论关联关系，供内部服务精确同步 Neo4j 使用"
      operationId: "getCommentsByIds"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchIdsDTO"
        required: true
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListCommentSyncVO"
...
  /comments/{id}:
    get:
      tags:
      - "评论模块"
      summary: "根据id查询评论"
      description: "根据id查询评论"
      operationId: "getCommentsById"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultCommentsVO"
    delete:
      tags:
      - "评论模块"
      summary: "删除评论"
      description: "根据id删除评论"
      operationId: "deleteComment"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
  /comments/user/{id}:
    get:
      tags:
      - "评论模块"
      summary: "根据用户id分页获取评论"
      description: "根据用户id分页获取评论"
      operationId: "listCommentsByUserId"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOCommentsVO"
  /comments/statistics/monthly-trend/{userId}:
    get:
      tags:
      - "评论模块"
      summary: "获取用户月度评论趋势（内部）"
      description: "获取用户本月评论的趋势，供内部服务远程调用"
      operationId: "getMonthlyCommentTrend"
      parameters:
      - name: "user_id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultMapDataVO"
  /comments/statistics/ai-count/{articleId}:
    get:
      tags:
      - "评论模块"
      summary: "获取文章AI评论数（内部）"
      description: "获取文章的AI评论数，供内部服务远程调用"
      operationId: "getAiCommentsNumByArticleId"
      parameters:
      - name: "article_id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultLong"
  /comments/neo4j-sync:
    get:
      tags:
      - "评论模块"
      summary: "获取评论表数据用于Neo4j同步（内部）"
      description: "获取评论表数据，支持增量同步，供FastAPI同步Neo4j使用"
      operationId: "getNeo4jSyncComments"
      parameters:
      - name: "updated_after"
        in: "query"
        required: false
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListCommentSyncVO"
  /comments/article/{id}:
    get:
      tags:
      - "评论模块"
      summary: "根据文章id分页获取评论"
      description: "根据文章id分页获取评论"
      operationId: "listCommentsByArticleId"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      - name: "sort_way"
        in: "query"
        required: false
        schema:
          type: "string"
          default: "create_time"
      - name: "size"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 10
      - name: "page"
        in: "query"
        required: false
        schema:
          type: "integer"
          format: "int32"
          default: 1
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOCommentsVO"
  /comments/article/ai/{id}:
    get:
      tags:
      - "评论模块"
      summary: "根据文章id获取AI评论"
      description: "根据文章id获取AI评论"
      operationId: "listAICommentsByArticleId"
      parameters:
      - name: "id"
        in: "path"
        required: true
        schema:
          type: "integer"
          format: "int64"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultListAICommentsVO"
  /comments/ai:
    get:
      tags:
      - "评论模块"
      summary: "获取AI评论信息"
      description: "分页获取AI评论信息列表，并支持AI类型和文章标题模糊查询"
      operationId: "listAIComments"
      parameters:
      - name: "page"
        in: "query"
        description: "页码"
        required: false
        schema:
          type: "string"
          description: "页码"
          example: 1
        example: 1
      - name: "size"
        in: "query"
        description: "每页条数"
        required: false
        schema:
          type: "string"
          description: "每页条数"
          example: 10
        example: 10
      - name: "username"
        in: "query"
        description: "用户名（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "用户名（模糊匹配）"
      - name: "article_title"
        in: "query"
        description: "文章标题（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "文章标题（模糊匹配）"
      - name: "content"
        in: "query"
        description: "评论内容（模糊匹配）"
        required: false
        schema:
          type: "string"
          description: "评论内容（模糊匹配）"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultPageVOCommentsVO"
...
  /comments/batch/{ids}:
    delete:
      tags:
      - "评论模块"
      summary: "批量删除评论"
      description: "根据id数组批量删除评论，多个id用英文逗号分隔"
      operationId: "deleteComments"
      parameters:
      - name: "ids"
        in: "path"
        required: true
        schema:
          type: "string"
      responses:
        "200":
          description: "OK"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ResultVoid"
...
components:
  schemas:
...
    ResultVoid:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "null"
          description: "返回数据"
          nullable: true
      description: "统一响应"
    CommentUpdateDTO:
      required:
      - "article_title"
      - "content"
      - "id"
      - "star"
      - "username"
      type: "object"
      properties:
        content:
          maxLength: 200
          minLength: 1
          type: "string"
          description: "评论内容"
        star:
          maximum: 10
          exclusiveMaximum: false
          minimum: 1
          exclusiveMinimum: false
          type: "number"
          description: "星级评分，1到10"
          format: "double"
        article_title:
          maxLength: 2147483647
          minLength: 1
          type: "string"
          description: "文章标题"
        username:
          maxLength: 2147483647
          minLength: 1
          type: "string"
          description: "评论用户名"
        id:
          minimum: 0
          type: "integer"
          description: "评论ID"
          format: "int32"
      description: "更新评论请求"
...
    ResultLong:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "integer"
          description: "返回数据"
          format: "int64"
      description: "统一响应"
...
    BatchIdsDTO:
      required:
      - "ids"
      type: "object"
      properties:
        ids:
          type: "array"
          description: "ID列表"
          items:
            type: "integer"
            description: "ID列表"
            format: "int64"
      description: "批量ID请求"
...
    CommentCreateDTO:
      required:
      - "article_title"
      - "content"
      - "star"
      - "username"
      type: "object"
      properties:
        content:
          maxLength: 200
          minLength: 1
          type: "string"
          description: "评论内容"
        star:
          maximum: 10
          exclusiveMaximum: false
          minimum: 1
          exclusiveMinimum: false
          type: "number"
          description: "星级评分，1到10"
          format: "double"
        article_title:
          maxLength: 2147483647
          minLength: 1
          type: "string"
          description: "文章标题"
        username:
          maxLength: 2147483647
          minLength: 1
          type: "string"
          description: "评论用户名"
      description: "创建评论请求"
    ArticleCommentScoresVO:
      type: "object"
      properties:
        article_id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        role_scores:
          type: "object"
          additionalProperties:
            $ref: "#/components/schemas/CommentScoreDTO"
          description: "按角色汇总的评分"
      description: "文章评论评分"
    CommentScoreDTO:
      type: "object"
      properties:
        average_score:
          type: "number"
          description: "平均评分"
          format: "double"
        count:
          type: "integer"
          description: "评论数量"
          format: "int64"
      description: "评论评分统计"
    ResultListArticleCommentScoresVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/ArticleCommentScoresVO"
      description: "统一响应"
    Comments:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "评论ID"
          format: "int64"
        content:
          type: "string"
          description: "评论内容"
        star:
          type: "number"
          description: "星级评分，1到10"
          format: "double"
        article_id:
          type: "integer"
          description: "文章ID"
          format: "int64"
        user_id:
          type: "integer"
          description: "用户ID"
          format: "int64"
        create_time:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_time:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "评论"
    CommentSyncVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "评论ID"
          format: "int64"
        user_id:
          type: "integer"
          description: "评论作者用户ID"
          format: "int64"
        article_id:
          type: "integer"
          description: "所属文章ID"
          format: "int64"
        star:
          type: "number"
          description: "评论评分，未评分时为空"
          format: "double"
        create_time:
          type: "string"
          description: "创建时间"
          format: "date-time"
        update_time:
          type: "string"
          description: "更新时间"
          format: "date-time"
      description: "评论同步数据"
    ResultListCommentSyncVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/CommentSyncVO"
      description: "统一响应"
...
    MapDataVO:
      type: "object"
      properties:
        data:
          type: "object"
          additionalProperties:
            type: "object"
            description: "动态键值对"
          description: "动态键值对"
      description: "键值数据"
    ResultMapDataVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/MapDataVO"
      description: "统一响应"
...
    CommentsVO:
      type: "object"
      properties:
        id:
          type: "integer"
          description: "评论ID"
          format: "int64"
        content:
          type: "string"
          description: "评论内容"
        star:
          type: "number"
          description: "星级评分"
          format: "double"
        username:
          type: "string"
          description: "用户名"
        pic:
          type: "string"
          description: "头像地址"
        article_title:
          type: "string"
          description: "文章标题"
        create_time:
          type: "string"
          description: "创建时间"
        update_time:
          type: "string"
          description: "更新时间"
      description: "评论视图"
    PageVOCommentsVO:
      type: "object"
      properties:
        total:
          type: "integer"
          description: "总记录数"
          format: "int64"
        list:
          type: "array"
          description: "当前页数据"
          items:
            $ref: "#/components/schemas/CommentsVO"
      description: "分页结果"
    ResultPageVOCommentsVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/PageVOCommentsVO"
      description: "统一响应"
    ResultCommentsVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          $ref: "#/components/schemas/CommentsVO"
      description: "统一响应"
    AICommentsVO:
      type: "object"
      properties:
        content:
          type: "string"
          description: "评论内容"
        star:
          type: "number"
          description: "星级评分"
          format: "double"
        ai_type:
          type: "string"
          description: "AI 类型"
        pic:
          type: "string"
          description: "头像地址"
      description: "AI 评论"
    ResultListAICommentsVO:
      type: "object"
      properties:
        code:
          type: "integer"
          description: "响应码"
          format: "int32"
        msg:
          type: "string"
          description: "响应信息"
        data:
          type: "array"
          description: "返回数据"
          items:
            $ref: "#/components/schemas/AICommentsVO"
      description: "统一响应"
...